requirements.txt
Flask
Flask-SQLAlchemy
pytest
pyjwt
flask-bcrypt